wip: userop status
