+++ packages/cli/.git-data.json
@@ -1,4 +1,4 @@
 {
   "branch": "websocket",
-  "commit": "edade258ebbdd7f84bc98b9584f1f73925018867"
+  "commit": "ba16529318b1d19bff87e92c5909abc5c6710885"
 }
