+++ packages/executor/src/services/BundlingService/relayers/kolibri.ts
@@ -102,7 +102,6 @@ export class KolibriRelayer extends BaseRelayer {
           await this.waitForEntries(entries).catch((err) =>
             this.logger.error(err, "Kolibri: Could not find transaction")
           );
-          await this.mempoolService.removeAll(entries);
         })
         .catch(async (err) => {
           this.reportFailedBundle();
