+++ packages/executor/src/services/BundlingService/relayers/merkle.ts
@@ -139,7 +139,6 @@ export class MerkleRelayer extends BaseRelayer {
           hash
         );
         await this.waitForTransaction(hash);
-        await this.mempoolService.removeAll(entries);
       } catch (err) {
         this.reportFailedBundle();
         await this.mempoolService.setStatus(entries, MempoolEntryStatus.New);
