+++ packages/executor/src/services/MempoolService/constants.ts
@@ -1,0 +1,2 @@
+export const MAX_MEMPOOL_USEROPS_PER_SENDER = 4;
+export const THROTTLED_ENTITY_MEMPOOL_COUNT = 4;
