+++ packages/types/src/executor/entities/MempoolEntry.ts
@@ -2,7 +2,8 @@ export enum MempoolEntryStatus {
   New = 0,
   Pending = 1,
   Submitted = 2,
-  IncludedToChain = 3,
+  OnChain = 3,
   Finalized = 4,
   Cancelled = 5,
+  Reverted = 6,
 }
